Added some first basic functions to manage dirty region/rectangles.
For now, the engine just keep track of updated region, and display them
(if feature and debug mode is activated). It will also allow me for now,
to ensure that ALL visible objects are correctly implemented (e.g.
extend me.Rect), and correctly returns the right coordinates depending
of the situation (for example, when scrolling, etc….).