Some experiments with WebGL (using the wrapper written by Corban Brook, Bobby Richter and Charles J. Cliffe)
If you to enable the feature and play with it, you will need : - a
browser supporting WebGL (but you should know that) - the
me.sys.enableWebGL flag set to true (in your main javascript) - the
WebGL wrapper library (to be loaded before melonJS), available here :
https://github.com/corbanbrook/webgl-2d  Current Limitations : -
context.measureText is not implemented in the wrapper, so be careful if
using system font (e.g. the melonJS text is not displayed during
loading) - globalAlpha is not supported (so if you change the opacity
settings for your layers/objects in Tiled, it won't work) -
context.scale & context.setTransform are apparently not working
correctly (so when sprites are flipped using the flip method, they are
not displayed).  I'm still hoping that the WebGL author will one day
update his library, as I don't know if it will be possible to get rid of
all the here above limitations, and since I don't want to maintain a
specific WebGL version in melonJS…. well I don't know actually, let's
keep this way for now, and let's see when we reach a stable version of
melonJS for the next public release.  Comments are welcome on this, and
feedback as well, if someone is reading this :)