I created two new properties for a texture's atlas: {boolean} hasTextureAnchorPoint - Whether or not the current frame defines its own anchor point. {me.Vector2d} anchorPoint - The current frame's anchor point (0,0) top-left to (1,1) bottom-right.
If you're creating a custom AnimationSheet, you can pass an anchorPoint in addition to the framewidth and frameheight.

var animationSheet = new me.AnimationSheet(0, 0, {
    image : "animationsheet",
    framewidth : 64,
    frameheight : 64,
    anchorPoint : new me.Vector2d(0.5, 0.5)
});

If you're using TexturePacker, frame anchorPoints are created automatically based on the "sourceSize", "spriteSourceSize", and "pivot" properties of the frame (assuming they all exist).

If an anchorPoint is set by the texture atlas json or manually like above, the entity.renderable's anchorPoint is drawn at the entity.body's anchorPoint. (as described in issue 715)

Otherwise, it acts the same as it does today, where the entity attempts to keep the renderable inside the body's bounds.