[#34][#724] few changes, see description
- changed the mapping logic, the api now reacts on the
`gamepadconnected` event and set the default remap/mapping at that
moment. This one need more love though as on OSX i’m noticing some
strange behaviour where the event is only triggered once when Chrome is
started, and not anymore when the page is reloaded. Is it the bug
reported by Jay ? If confirmed not a bug in my code, we should probably
revert this back in the update loop (just add a test if `mapped[index]`
is defined)

- added the `setGamepadMapping` function, but without though the
`normalize_fn` stuff