Clarification of body.Velocity Comments
The previous description was very confusing.  I eliminated the following line because it just doesn't make any sense.
eliminated --> * note : velocity is by default limited to the same value, see

In the future it would be useful to state how the accel / vel members of the body are intended to be used.