Improvements in Documenting how the body physics work
The Movement and Utilization of the Body.update call was very difficult to follow.  In part because there are were calls to private functions that were not explained in the documentation.  I believe my descriptions more accurately represent the intended approach.
There are 2 Issues still unaddressed in this commit  1) the callback 'onBodyUpdate'  seems inconsistently placed.  If it is not to be called after Body.Update then perhaps the name is incorrect.  2) The accel property updates the maxVelocity... so its not clear the purpose of this property.  I would think that like regular physics  vel = 1/2 a t^2 + v_(@time 0).