Corrected incorrect pointer offset - Please check - Critical
This is a potentially critical problem.  The click was off on an UI click events.. the previous setting this.pos.x  registered to the incorrect location.  I could not exactly nail down which coordinate this.pos.x.  I think it was the clientX which seemed to be consistently different from the canvas element position which is given by either event.layerX/Y  or event.offsetX/Y... So it seems that the clientXY is different from the canvas application position.  I tested this in my application and it seems to work well.