[#988] fix implementation and usage of the isDirty flag across Renderables
isDirty is now a public property of the me.Renderable object, that user can force if necessary, and used to define if a renderable needs to be redrawn (will be set to true internally when a renderable state is changing e.g. transform, animation ... and set to false automatically at the end of the draw loop.